refactor(use-if-elif-else): move some relevant code into separate file - z3_block_analysis.py and add some documentation
