+++ edulint/linting/checkers/simplifiable_if.py
@@ -4,7 +4,7 @@ from astroid import extract_node
 from typing import TYPE_CHECKING, Optional, Tuple, Union, List, Any, Dict, Set
 from enum import Enum
 
-from z3 import ArithRef, ExprRef, BoolRef, And, Or, Not, If, Implies, is_bool, BoolVal
+from z3 import ArithRef, ExprRef, BoolRef, And, Or, Not, BoolVal
 
 from pylint.checkers import BaseChecker  # type: ignore
 from pylint.checkers.utils import only_required_for_messages
@@ -18,9 +18,13 @@ from edulint.linting.checkers.z3_analysis import (
     initialize_variables,
     convert_condition_to_z3_expression,
     unsatisfiable,
-    create_prefixed_var,
     _is_bool_node,
-    convert_to_bool,
+)
+
+from edulint.linting.checkers.z3_block_analysis import (
+    END_NODES,
+    validate_and_initialize_variables_for_Z3_block_analysis,
+    convert_conditions_with_blocks_after_each_to_Z3,
 )
 
 from edulint.linting.checkers.utils import (
@@ -33,9 +37,6 @@ from edulint.linting.checkers.utils import (
     is_negation,
     is_parents_elif,
     if_elif_has_else_block,
-    is_chained_assignment,
-    has_more_assign_targets,
-    get_assign_targets,
 )
 
 from edulint.linting.analyses.cfg.utils import syntactic_children_locs
@@ -1101,155 +1102,10 @@ class SimplifiableIf(BaseChecker):  # type: ignore
 
         return consecutive_ifs
 
-    ALLOWED_EXPR_NODES_FOR_Z3_BLOCK_ANALYSIS = (
-        nodes.BinOp,
-        nodes.BoolOp,
-        nodes.Call,
-        nodes.Compare,
-        nodes.Const,
-        nodes.Name,
-        nodes.Tuple,
-        nodes.UnaryOp,
-    )
-
-    END_NODES = (
-        nodes.Break,
-        nodes.Continue,
-        nodes.Raise,
-        nodes.Return,
-    )
-
-    ALWAYS_PURE_ALLOWED_NODES_FOR_Z3_BLOCK_ANALYSIS = (
-        *END_NODES,
-        nodes.Pass,
-    )
-
-    ALLOWED_NODES_FOR_Z3_BLOCK_ANALYSIS = (
-        *ALLOWED_EXPR_NODES_FOR_Z3_BLOCK_ANALYSIS,
-        *ALWAYS_PURE_ALLOWED_NODES_FOR_Z3_BLOCK_ANALYSIS,
-        nodes.Assign,
-        nodes.AnnAssign,
-        nodes.AugAssign,
-        nodes.Expr,
-        nodes.Assert,
-        nodes.IfExp,
-        nodes.If,
-    )
-
-    def _check_purity_for_Z3_block_analysis(self, node: nodes.NodeNG) -> bool:
-        if isinstance(node, nodes.Assert):
-            return is_pure_expression(node.test)
-
-        if isinstance(node, self.ALLOWED_EXPR_NODES_FOR_Z3_BLOCK_ANALYSIS):
-            return is_pure_expression(node)
-
-        if isinstance(node, self.ALWAYS_PURE_ALLOWED_NODES_FOR_Z3_BLOCK_ANALYSIS):
-            return True
-
-        if isinstance(node, (nodes.Assign, nodes.AnnAssign, nodes.AugAssign, nodes.Expr)):
-            return self._check_purity_for_Z3_block_analysis(node.value)
-
-        # not doing nested IfExps
-        if isinstance(node, nodes.IfExp):
-            return (
-                is_pure_expression(node.test)
-                and is_pure_expression(node.body)
-                and is_pure_expression(node.orelse)
-            )
-
-        return False
-
-    def is_assignment(self, node: nodes.NodeNG) -> bool:
-        return isinstance(node, (nodes.Assign, nodes.AnnAssign, nodes.AugAssign))
-
     def _vars_from_node_are_modified_in(
         self, node: nodes.NodeNG, nodes: List[nodes.NodeNG]
     ) -> bool:
         return modified_in(list(vars_in(node).keys()), nodes)
-
-    def _vars_from_non_linear_arithmetic_are_modified_in(
-        self, node: nodes.NodeNG, nodes: List[nodes.NodeNG]
-    ) -> bool:
-        """
-        Note that could be enhanced by checking only the variables from expressions of this node
-        that are non-linear arithmetic (given by function _is_expression_with_nonlinear_arithmetic()
-        from z3_analysis) or that are a function call (not including abs()).
-
-        Here we check whether any variable from this node is modified in `nodes` for simplification,
-        most likely it will be only the two variables anyway, like in expressions `m % n == 0`, ...
-        """
-        return modified_in(list(vars_in(node).keys()), nodes)
-
-    def _allowed_node_for_Z3_block_analysis(self, node: nodes.NodeNG) -> bool:
-        return (
-            isinstance(node, self.ALLOWED_NODES_FOR_Z3_BLOCK_ANALYSIS)
-            and not is_chained_assignment(node)
-            and (not isinstance(node, nodes.IfExp) or self.is_assignment(node.parent))
-        )
-
-    def _initialize_variables_in_node(
-        self,
-        node: nodes.NodeNG,
-        context: List[nodes.NodeNG],
-        initialized_variables: Dict[str, ArithRef],
-    ) -> bool:
-        if isinstance(node, nodes.Assert):
-            node = node.test
-
-        nodes_tmp = [node]
-
-        if isinstance(node, nodes.Assign) and has_more_assign_targets(node):
-            nodes_tmp = list(node.value.get_children())
-        elif self.is_assignment(node):
-            nodes_tmp = [node.value]
-
-        nodes_for_initialization: List[nodes.NodeNG] = []
-        for node in nodes_tmp:
-            if isinstance(node, nodes.IfExp):
-                nodes_for_initialization.extend([node.test, node.body, node.orelse])
-            else:
-                nodes_for_initialization.append(node)
-
-        for current_node in nodes_for_initialization:
-            if not isinstance(current_node, self.ALLOWED_EXPR_NODES_FOR_Z3_BLOCK_ANALYSIS):
-                return False
-            # because the variables from `m%n == 0` for example are modified later we cannot take `m%n` as
-            # a variable, because it would have different value later then now.
-            dont_make_up_new_vars = self._vars_from_non_linear_arithmetic_are_modified_in(
-                current_node, context
-            )
-            if not initialize_variables(
-                current_node, initialized_variables, False, None, dont_make_up_new_vars
-            ):
-                return False
-
-        return True
-
-    def _validate_and_initialize_variables_for_Z3_block_analysis(
-        self,
-        node: nodes.NodeNG,
-        initialized_variables: Dict[str, ArithRef],
-        context: List[nodes.NodeNG],
-    ) -> bool:
-        if may_contain_mutable_var(node):
-            return False
-
-        for loc in syntactic_children_locs(node):
-            if not (
-                self._allowed_node_for_Z3_block_analysis(loc.node)
-                and self._check_purity_for_Z3_block_analysis(loc.node)
-            ) or (isinstance(loc.node, self.END_NODES) and loc.node.parent is node):
-                return False
-
-            if isinstance(
-                loc.node, self.ALWAYS_PURE_ALLOWED_NODES_FOR_Z3_BLOCK_ANALYSIS
-            ) or isinstance(loc.node, nodes.Expr):
-                continue
-
-            if not self._initialize_variables_in_node(loc.node, context, initialized_variables):
-                return False
-
-        return True
 
     def _validate_and_initialize_variables_for_use_if_elif_else(
         self, consecutive_ifs: List[nodes.If]
@@ -1265,7 +1121,7 @@ class SimplifiableIf(BaseChecker):  # type: ignore
         for i, if_stmt in enumerate(consecutive_ifs):
             if not if_elif_has_else_block(
                 if_stmt
-            ) and self._validate_and_initialize_variables_for_Z3_block_analysis(
+            ) and validate_and_initialize_variables_for_Z3_block_analysis(
                 if_stmt,
                 initialized_variables,
                 consecutive_ifs[i:],
@@ -1294,422 +1150,6 @@ class SimplifiableIf(BaseChecker):  # type: ignore
             ifs_with_initialized_vars.append((current_ifs, initialized_variables))
 
         return ifs_with_initialized_vars, not_valid_consecutive_ifs
-
-    def _change_assertions_and_vars_after_elif_block(
-        self,
-        if_conditions: List[ExprRef],
-        return_encountered: bool,
-        new_assertions: List[ExprRef],
-        var_changes: Dict[str, ArithRef],
-        assertions: List[ExprRef],
-        current_condition: ExprRef,
-        changed_vars_in_any_branch: Set[str],
-        var_changes_in_each_branch: List[Dict[str, ArithRef]],
-    ):
-        previous_conditions_negated = And([Not(cond) for cond in if_conditions])
-
-        if return_encountered:
-            assertions.append(
-                Implies(previous_conditions_negated, Not(current_condition))
-                if len(if_conditions) > 0
-                else Not(current_condition)
-            )
-        else:
-            for assertion in new_assertions:
-                assertions.append(
-                    Implies(
-                        (
-                            And(*previous_conditions_negated, current_condition)
-                            if len(if_conditions) > 0
-                            else current_condition
-                        ),
-                        assertion,
-                    )
-                )
-
-            for var_name in var_changes.keys():
-                changed_vars_in_any_branch.add(var_name)
-
-            var_changes_in_each_branch.append((var_changes, return_encountered))
-
-            if_conditions.append(current_condition)
-
-    def _how_var_changed_after_if(
-        self,
-        var_name: str,
-        if_conditions: List[ExprRef],
-        var_changes_in_each_branch: List[Tuple[Dict[str, ArithRef], bool]],
-        var_value_before_if: ArithRef,
-        current_block: int,
-        has_else_block: bool,
-    ) -> ExprRef:
-        if current_block >= len(var_changes_in_each_branch) and not has_else_block:
-            return var_value_before_if
-
-        var_changes, return_found = var_changes_in_each_branch[current_block]
-
-        if current_block == len(var_changes_in_each_branch) - 1 and has_else_block:
-            return var_changes.get(var_name, var_value_before_if)
-
-        if return_found:
-            return self._how_var_changed_after_if(
-                var_name,
-                if_conditions,
-                var_changes_in_each_branch,
-                var_value_before_if,
-                current_block + 1,
-                has_else_block,
-            )
-
-        return If(
-            if_conditions[current_block],
-            var_changes.get(var_name, var_value_before_if),
-            self._how_var_changed_after_if(
-                var_name,
-                if_conditions,
-                var_changes_in_each_branch,
-                var_value_before_if,
-                current_block + 1,
-                has_else_block,
-            ),
-        )
-
-    def update_relations_between_vars(
-        self,
-        var_name: str,
-        var_value: ExprRef,
-        var_rewrite_counts: Dict[str, int],
-        accumulated_relations_between_vars: List[ExprRef],
-        initialized_variables: Dict[str, ArithRef],
-    ) -> ArithRef:
-        var_rewrite_counts[var_name] += 1
-        prefix = str(var_rewrite_counts[var_name])
-
-        var = create_prefixed_var(prefix, initialized_variables[var_name], var_name)
-
-        accumulated_relations_between_vars.append(
-            (convert_to_bool(var) if is_bool(var_value) else var) == var_value
-        )
-
-        return var
-
-    def _create_new_var_after_if(
-        self,
-        var_name: str,
-        if_conditions: List[ExprRef],
-        var_changes_in_each_branch: List[Tuple[Dict[str, ArithRef], bool]],
-        accumulated_relations_between_vars: List[ExprRef],
-        var_rewrite_counts: Dict[str, int],
-        initialized_variables: Dict[str, ArithRef],
-        has_else_block: bool,
-    ) -> ArithRef:
-        new_var_value = self._how_var_changed_after_if(
-            var_name,
-            if_conditions,
-            var_changes_in_each_branch,
-            initialized_variables[var_name],
-            0,
-            has_else_block,
-        )
-
-        return self.update_relations_between_vars(
-            var_name,
-            new_var_value,
-            var_rewrite_counts,
-            accumulated_relations_between_vars,
-            initialized_variables,
-        )
-
-    def _changed_vars_after_if(
-        self,
-        node: nodes.If,
-        initialized_variables: Dict[str, ArithRef],
-        accumulated_relations_between_vars: List[ExprRef],
-        var_rewrite_counts: Dict[str, int],
-    ) -> Optional[Tuple[Dict[str, ArithRef], List[ExprRef], bool]]:
-        new_vars: Dict[str, ArithRef] = {}
-        assertions: List[ExprRef] = []
-        changed_vars_in_any_branch: Set[str] = set()
-
-        always_returns = True
-        if_conditions: List[ExprRef] = []
-        var_changes_in_each_branch: List[Tuple[Dict[str, ArithRef], bool]] = []
-        current_node = node
-        while True:
-            current_condition = convert_condition_to_z3_expression(
-                current_node.test, initialized_variables, None
-            )[0]
-            after_block = self._changed_vars_after_block(
-                current_node.body,
-                initialized_variables.copy(),
-                accumulated_relations_between_vars,
-                var_rewrite_counts,
-            )
-
-            if current_condition is None or after_block is None:
-                return None
-
-            var_changes, new_assertions, return_encountered = after_block
-            always_returns = always_returns and return_encountered
-
-            self._change_assertions_and_vars_after_elif_block(
-                if_conditions,
-                return_encountered,
-                new_assertions,
-                var_changes,
-                assertions,
-                current_condition,
-                changed_vars_in_any_branch,
-                var_changes_in_each_branch,
-            )
-
-            if not current_node.has_elif_block():
-                break
-
-            current_node = current_node.orelse[0]
-
-        has_else_block = False
-        if len(current_node.orelse) > 0:
-            after_block = self._changed_vars_after_block(
-                current_node.orelse,
-                initialized_variables.copy(),
-                accumulated_relations_between_vars,
-                var_rewrite_counts,
-            )
-
-            if after_block is None:
-                return None
-
-            var_changes, new_assertions, return_encountered = after_block
-            always_returns = always_returns and return_encountered
-
-            if return_encountered:
-                # this is negation of And([Not(cond) for cond in if_conditions]) (ie condition that holds when we get to else)
-                assertions.append(Or(if_conditions))
-            else:
-                for assertion in new_assertions:
-                    assertions.append(
-                        Implies(And([Not(cond) for cond in if_conditions]), assertion)
-                    )
-
-                for var_name in var_changes.keys():
-                    changed_vars_in_any_branch.add(var_name)
-
-                var_changes_in_each_branch.append((var_changes, return_encountered))
-
-            has_else_block = True
-        else:
-            always_returns = False
-
-        if always_returns:
-            return {}, [], True
-
-        i = len(var_changes_in_each_branch) - 1
-        while i >= 0 and var_changes_in_each_branch[i][1]:
-            var_changes_in_each_branch.pop()
-            i -= 1
-
-        for var_name in changed_vars_in_any_branch:
-            new_vars[var_name] = self._create_new_var_after_if(
-                var_name,
-                if_conditions,
-                var_changes_in_each_branch,
-                accumulated_relations_between_vars,
-                var_rewrite_counts,
-                initialized_variables,
-                has_else_block,
-            )
-
-        return new_vars, assertions, False
-
-    def _convert_expression_in_assignment_to_Z3(
-        self, node: nodes.NodeNG, initialized_variables: Dict[str, ArithRef]
-    ) -> Optional[ExprRef]:
-        if isinstance(node, nodes.IfExp):
-            test = convert_condition_to_z3_expression(node.test, initialized_variables, None)[0]
-            body = convert_condition_to_z3_expression(
-                node.body, initialized_variables, node.parent
-            )[0]
-            orelse = convert_condition_to_z3_expression(
-                node.orelse, initialized_variables, node.parent
-            )[0]
-
-            if test is None or body is None or orelse is None:
-                return None
-
-            return If(test, body, orelse)
-
-        return convert_condition_to_z3_expression(node, initialized_variables, node.parent)[0]
-
-    def _get_assigned_expression_in_AugAssign(self, node: nodes.AugAssign) -> nodes.NodeNG:
-        assigned_expression = nodes.BinOp(
-            op=node.op[:-1],
-            lineno=node.lineno,
-            col_offset=node.col_offset,
-            parent=node,
-            end_lineno=node.end_lineno,
-            end_col_offset=node.end_col_offset,
-        )
-
-        if not isinstance(node.target, nodes.AssignName):
-            return assigned_expression
-
-        assigned_expression.postinit(
-            nodes.Name(
-                name=node.target.name,
-                lineno=node.target.lineno,
-                col_offset=node.target.col_offset,
-                parent=assigned_expression,
-                end_lineno=node.target.end_lineno,
-                end_col_offset=node.target.end_col_offset,
-            ),
-            node.value,
-        )
-        return assigned_expression
-
-    def _update_vars_after_assignment(
-        self,
-        assignment: Union[nodes.Assign, nodes.AnnAssign, nodes.AugAssign],
-        initialized_variables: Dict[str, ArithRef],
-        accumulated_relations_between_vars: List[ExprRef],
-        var_rewrite_counts: Dict[str, int],
-    ) -> Optional[Dict[str, ArithRef]]:
-        if isinstance(assignment, nodes.Assign) and has_more_assign_targets(assignment):
-            values = list(assignment.value.get_children())
-        elif isinstance(assignment, nodes.AugAssign):
-            values = [self._get_assigned_expression_in_AugAssign(assignment)]
-        else:
-            values = [assignment.value]
-
-        targets = get_assign_targets(assignment)
-
-        new_vars: Dict[str, ArithRef] = {}
-
-        for i, target in enumerate(targets):
-            if not isinstance(target, nodes.AssignName):
-                return None
-
-            if target.name not in initialized_variables:
-                continue
-
-            converted = self._convert_expression_in_assignment_to_Z3(
-                values[i], initialized_variables
-            )
-
-            if converted is None:
-                return None
-
-            var = self.update_relations_between_vars(
-                target.name,
-                converted,
-                var_rewrite_counts,
-                accumulated_relations_between_vars,
-                initialized_variables,
-            )
-
-            new_vars[target.name] = var
-
-        initialized_variables.update(new_vars)
-        return new_vars
-
-    def _changed_vars_after_block(
-        self,
-        block: List[nodes.NodeNG],
-        initialized_variables: Dict[str, ArithRef],
-        accumulated_relations_between_vars: List[ExprRef],
-        var_rewrite_counts: Dict[str, int],
-    ) -> Optional[Tuple[Dict[str, ArithRef], List[ExprRef], bool]]:
-        new_vars: Dict[str, ArithRef] = {}
-        assertions: List[ExprRef] = []
-
-        for node in block:
-            if self.is_assignment(node):
-                updated_vars = self._update_vars_after_assignment(
-                    node,
-                    initialized_variables,
-                    accumulated_relations_between_vars,
-                    var_rewrite_counts,
-                )
-
-                if updated_vars is None:
-                    return None
-
-                new_vars.update(updated_vars)
-            elif isinstance(node, self.END_NODES):
-                return ({}, [], True)
-            elif isinstance(node, nodes.If):
-                after_if = self._changed_vars_after_if(
-                    node,
-                    initialized_variables,
-                    accumulated_relations_between_vars,
-                    var_rewrite_counts,
-                )
-
-                if after_if is None:
-                    return None
-
-                var_changes, new_assertions, always_returns = after_if
-
-                if always_returns:
-                    return ({}, [], True)
-
-                new_vars.update(var_changes)
-                initialized_variables.update(var_changes)
-                assertions.extend(new_assertions)
-            elif isinstance(node, nodes.Assert):
-                converted = convert_condition_to_z3_expression(
-                    node.test, initialized_variables, None
-                )[0]
-                if converted is None:
-                    return None
-
-                assertions.append(converted)
-
-        return new_vars, assertions, False
-
-    def convert_conditions_with_blocks_after_each_to_Z3(
-        self,
-        conditions: List[List[nodes.NodeNG]],
-        blocks: List[List[nodes.NodeNG]],
-        initialized_variables: Dict[str, ArithRef],
-    ) -> Tuple[List[ExprRef], List[List[ExprRef]]]:
-        """
-        `conditions[i]` is a list of conditions and `blocks[i]` is a block (list of nodes) between the conditions
-        on position `i` and conditions on position `i + 1`. So it must hold that len(conditions) == len(blocks) + 1
-        """
-        assert len(conditions) == len(blocks) + 1
-
-        accumulated_relations_between_vars: List[ExprRef] = []
-        converted_conditions: List[List[ExprRef]] = []
-
-        var_rewrite_counts = {var: 0 for var in initialized_variables}
-
-        for i, conds in enumerate(conditions):
-            converted_conditions.append([])
-            for cond in conds:
-                converted = convert_condition_to_z3_expression(cond, initialized_variables, None)[0]
-                if converted is None:
-                    return ([], [])
-
-                converted_conditions[i].append(converted)
-
-            if i < len(blocks):
-                after_block = self._changed_vars_after_block(
-                    blocks[i],
-                    initialized_variables.copy(),
-                    accumulated_relations_between_vars,
-                    var_rewrite_counts,
-                )
-
-                if after_block is None:
-                    return ([], [])
-
-                new_vars, assertions, _ = after_block
-
-                initialized_variables.update(new_vars)
-                accumulated_relations_between_vars.extend(assertions)
-
-        return accumulated_relations_between_vars, converted_conditions
 
     def _get_list_of_test_conditions(self, node: nodes.If) -> List[nodes.NodeNG]:
         test_conditions: List[nodes.NodeNG] = []
@@ -1771,12 +1211,12 @@ class SimplifiableIf(BaseChecker):  # type: ignore
         self,
         ifs: List[nodes.If],
         converted_conditions: List[List[ExprRef]],
-        relations_between_vars_in_Z3: BoolRef,
+        relations_between_vars: BoolRef,
     ) -> None:
         i = 0
         while i < len(converted_conditions):
             first_unmergable_index = self._index_of_first_unmergable_consecutive_if(
-                relations_between_vars_in_Z3, converted_conditions, i
+                relations_between_vars, converted_conditions, i
             )
 
             if first_unmergable_index > i + 1:
@@ -1796,20 +1236,18 @@ class SimplifiableIf(BaseChecker):  # type: ignore
         self, ifs: List[nodes.If], initialized_variables: Dict[str, ArithRef]
     ) -> None:
         relations_between_vars, converted_conditions = (
-            self.convert_conditions_with_blocks_after_each_to_Z3(
+            convert_conditions_with_blocks_after_each_to_Z3(
                 [self._get_list_of_test_conditions(if_node) for if_node in ifs],
                 [[if_stmt] for if_stmt in ifs[:-1]],
                 initialized_variables,
             )
         )
 
-        relations_between_vars_in_Z3 = And(relations_between_vars)
-
-        self._find_mergable_consecutive_ifs(ifs, converted_conditions, relations_between_vars_in_Z3)
+        self._find_mergable_consecutive_ifs(ifs, converted_conditions, relations_between_vars)
 
     def _if_elif_has_end_node_in_any_branch(self, node: nodes.If) -> bool:
         for loc in syntactic_children_locs(node):
-            if isinstance(loc.node, self.END_NODES):
+            if isinstance(loc.node, END_NODES):
                 return True
 
         return False
